refactor(web): consolidate core store ownership
Move learning and AI state into architectureStore, move promotion flow state into uiStore, and stop direct cross-store writes in store actions.

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
