+++ apps/web/src/entities/store/slices/helpers.ts
@@ -9,7 +9,6 @@ import {
   resetHistory,
 } from '../../../shared/utils/history';
 import { generateId } from '../../../shared/utils/id';
-import { useUIStore } from '../uiStore';
 import type { ArchitectureState } from './types';
 
 const DEFAULT_WORKSPACE_NAME = 'My Architecture';
@@ -114,7 +113,6 @@ export function resetTransientState(): Pick<
   ArchitectureState,
   'validationResult' | 'history' | 'canUndo' | 'canRedo'
 > {
-  useUIStore.getState().setDiffMode(false);
   return {
     validationResult: null,
     history: resetHistory(),
